requirements/nox-requirements.txt
nox
GitPython
python-dotenv